+++ app/Filament/Resources/Companies/Tables/CompaniesTable.php
@@ -97,6 +97,7 @@ class CompaniesTable
                 ActionGroup::make([
                     CompanyBackupActions::scheduleSettings(),
                     CompanyBackupActions::runNow(),
+                    CompanyBackupActions::downloadNow(),
                     CompanyBackupActions::export(),
                     CompanyBackupActions::importInto(),
                     CompanyBackupActions::wipe(),